pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[project]
name = "agent-delegate"
version = "0.2.0"
description = "Multi-agent orchestration system using OpenHands SDK"
readme = "CLAUDE.md"
requires-python = ">=3.10"
authors = [
    {name = "Agent-Delegate Project"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openhands-sdk>=1.8.1",
    "openhands-tools",
    "PyYAML>=6.0.0",
    "jsonschema>=4.26.0",
    "click>=8.3.0",
    "pydantic>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "black>=24.8.0",
    "ruff>=0.7.0",
    "mypy>=1.11.0",
]

[project.scripts]
agent-delegate = "orchestrator.cli:main"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
